Moved FlushMode to Commands.ts
