+++ node/src/GlideClusterClient.ts
@@ -10,6 +10,7 @@ import {
     ReturnType,
 } from "./BaseClient";
 import {
+    FlushMode,
     InfoOptions,
     LolwutOptions,
     createClientGetName,
@@ -32,7 +33,6 @@ import {
     createPublish,
     createTime,
 } from "./Commands";
-import { FlushMode } from "./commands/FlushMode";
 import { RequestError } from "./Errors";
 import { command_request, connection_request } from "./ProtobufMessage";
 import { ClusterTransaction } from "./Transaction";
